requirements.txt
fastapi
uvicorn
python-multipart
llama-index-core
llama-index-llms-llama-cpp
llama-index-embeddings-huggingface
llama-index-readers-file
llama-index-vector-stores-postgres
llama-index-llms-google-genai
sentence-transformers
pydantic-settings
python-dotenv
pypdf
docx2txt
hf_xet
python-docx
sqlalchemy
pgvector
alembic
psycopg2-binary
asyncpg
pandas
torch